:wrench: move helpers to utils.js / remove unused packages
