+++ package.json
@@ -25,16 +25,12 @@
     "express": "^4.16.4",
     "express-session": "^1.15.6",
     "jquery": "^3.3.1",
-    "js-yaml": "^3.12.0",
-    "marked": "^0.5.1",
     "mime": "^2.3.1",
-    "mkdirp": "^0.5.1",
     "moment": "^2.22.2",
     "multer": "^1.4.1",
     "nodemailer": "^4.6.8",
     "pg": "^7.6.0",
     "request": "^2.88.0",
-    "sharp": "^0.21.0",
     "socket.io": "^2.1.1",
     "source-sans-pro": "^2.40.0"
   },
